+++ lib/src/controller/profile_controller.dart
@@ -9,8 +9,8 @@ class ProfileController extends GetxController {
   Rx<File?> selectedImage = Rx<File?>(null);
   final picker = ImagePicker();
 
-  void pfmgo() {
-    Get.to(() => const ProfileModify());
+  void pfmgo(ProfileController controller) {
+    Get.to(() => ProfileModify(controller));
   }
 
   Future<void> pickImage() async {
@@ -19,7 +19,7 @@ class ProfileController extends GetxController {
     if (pickedFile != null) {
       profileImagePath.value = pickedFile.path;
     } else {
-      print('사진을 서택하지 않았습니다.');
+      // print('사진을 서택하지 않았습니다.');
     }
   }
 }
